Minor: consolidate some parquet_testing examples into arrow_reader
